+++ src/styles/blogPost.module.css
@@ -1,0 +1,28 @@
+.blogPost {
+  width: 320px;
+}
+
+.blogPost>div {
+  width: 100%;
+  border-radius: 5px;
+  overflow: hidden;
+}
+
+.blogPost>p {
+  margin-top: 10px;
+  font-size: 14px;
+  color: var(--gray);
+}
+
+.blogPost>a {
+  color: var(--blue-dark);
+}
+
+.blogPost>a:hover {
+  text-decoration: underline;
+}
+
+.blogPost h3 {
+  font-size: 24px;
+  font-weight: 600;
+}
