+++ src/styles/home/blog.module.css
@@ -11,35 +11,6 @@
   justify-content: center;
 }
 
-.blogPost {
-  width: 320px;
-}
-
-.blogPost>div {
-  width: 100%;
-  border-radius: 5px;
-  overflow: hidden;
-}
-
-.blogPost>p {
-  margin-top: 10px;
-  font-size: 14px;
-  color: var(--gray);
-}
-
-.blogPost>a {
-  color: var(--blue-dark);
-}
-
-.blogPost>a:hover {
-  text-decoration: underline;
-}
-
-.blogPost h3 {
-  font-size: 24px;
-  font-weight: 600;
-}
-
 .btnContainer {
   margin-top: 70px;
   text-align: center;
